Hackfix for DST changes in Firefox
`toArray('months')` gives 13 months for some years because of a bug in
Firefox and/or momentjs. This is because if the DST change is at
midnight adding 24h to a day won't get you the next date and this is
what momentjs is doing behind the scenes.

This makes these DST changes not affect the date.