Fix calendar jumping on day select
The problem is that when you select a new date it will call
`props.onSelect`, then `componentWillReceiveProps` will notice that the
incoming `selectedDate` has changed and decides to scroll to it. This is
the correct behaviour **only** when we specify an **initial date** but
this the date we just selected.

This fix compares the incoming selected date against the internal
component state. Also `onSelect` is now called after the internal state
has been set. Otherwise this woudln't work.

The problem is that now we can't pass the event to the callback, since
react recycles those objects by default. This feature is undocumented
though so maybe yolo?