Add 'adjustRangeFunc' prop for withRange HOC
When adjustRangeFunc is a valid function, it's called with arguments
(selectionStart, end) in "YYYY-MM-DD" string format for every attempt to
end a selection range or while hovering (after 1st start click) to
modify a selection range.

adjustRangeFunc can return an Object with up to three recognized keys:
`start`, `end`, and `startOverride`.

`start` and `end` override the date range that will be committed.
It does not matter whether one is earlier than the other.

`startOverride`, if provided, overrides what date the Calendar thinks
the date selection started at. Usually it's unnecessary to use this.

If properties are missing in the returned object, a best guess is made.

To not modify the date range, return either {start: selectionStart, end:
date} or null.

This is work for https://github.com/toggl/webapp/issues/7091